Remove getMethods() methods, b/c they're unused
The remote plugins already have a reflection method that lists all
public methods. I added a comment to explain that. There doesn't seem to
be a use case for public methods on a remote component that are not part
of the XMLRPC API.

There also doesn't seem to be a use case for the getMethods method in
helper plugins.